pyproject.toml
[project]
name = "python-template"
version = "0.1.0"
description = ""
authors = ["NABORAJ SARKAR"]
requires-python = ">=3.11"
dependencies = [
    "pygithub>=2.8.1",
    "python-dotenv>=1.2.1",
    "python-telegram-bot>=22.6",
    "schedule>=1.2.2",
]
